feat: change alergicFoods to intolerances and add enum-based intolerance support

+++ src/main/java/com/gdgswu/planeat/domain/food/Food.java
@@ -22,7 +22,7 @@ public class Food {
     private Double cost;
     private String costUnit;
 
-    private Integer calories;
+    private Double calories;
     private Double carbs;
     private Double protein;
     private Double fat;
